Communication between mcu and backend almost finished.
