+++ online_module/DP_System-Back/src/main/java/dp/system/back/models/ParkingLot.java
@@ -2,7 +2,6 @@ package dp.system.back.models;
 
 import java.util.List;
 
-
 public class ParkingLot {
 
     private List<Integer> parkingSpaces;
